requirements.txt
requests
PyGithub
openai
python-dotenv
